// frontend/src/backend.d.ts
export interface DiscountCode {
    id: string;
    active: boolean;
    value: number;
    code: string;
    createdAt: bigint;
    discountType: Variant_percentage_fixedAmount;
}
// ...
export interface ShoppingItem {
    productName: string;
    currency: string;
    quantity: bigint;
    priceInCents: bigint;
    productDescription: string;
}
// ...
export enum Variant_percentage_fixedAmount {
    percentage = "percentage",
    fixedAmount = "fixedAmount"
}